♻️ Move pipelines to src/usecases, selectivity to policy (Task_3)
Use-case pipelines leave the internal/repositories catch-all:
remember, link, retrieve, write_planning, reconciliation, and
correct_forget (kept together — shared cascade/provenance internals
dominate). Plan-construction logic moves from the public api layer
into usecases/write_planning per ADR-I-0012; the api helper file
remains as a re-export shim. retrieval_selectivity joins policy/.
