+++ src/usecases.rs
@@ -1,0 +1,21 @@
+pub(crate) mod correct_forget;
+pub(crate) mod link;
+pub(crate) mod reconciliation;
+pub(crate) mod remember;
+pub(crate) mod retrieve;
+pub(crate) mod write_planning;
+
+pub(crate) use correct_forget::CorrectionForgetPipeline;
+pub(crate) use link::{admit_link, LinkAdmissionDecision, LinkAdmissionEvidence, LinkPipeline};
+pub(crate) use reconciliation::{
+    reconcile_graph_vector_stores, ReconciliationDiagnostic, ReconciliationDriftKind,
+    ReconciliationReport,
+};
+pub(crate) use remember::{
+    RememberPipeline, RememberPipelineDraft, RememberPipelineOutcome, VectorIndexingFailure,
+};
+pub(crate) use retrieve::RetrievePipeline;
+pub(crate) use write_planning::{
+    WritePlanCommitValues, WritePlanValidationDecision, WritePlanValidationVerdict,
+    WritePlanValidator,
+};
